# crates/barbacane-control/openapi.yaml
openapi: 3.1.0
info:
  title: Barbacane Control Plane API
  description: |
    REST API for managing API specifications, plugins, artifacts, and compilations
    in the Barbacane API Gateway control plane.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090
    description: Local development server
...
paths:
...
  /plugins/{name}:
    get:
      tags: [Plugins]
      summary: List plugin versions
      description: List all versions of a specific plugin.
      operationId: listPluginVersions
      parameters:
        - $ref: '#/components/parameters/pluginName'
      responses:
        '200':
          description: List of plugin versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plugin'
        '404':
          description: Plugin not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /plugins/{name}/{version}:
    get:
      tags: [Plugins]
      summary: Get plugin
      description: Get metadata for a specific plugin version.
      operationId: getPlugin
      parameters:
        - $ref: '#/components/parameters/pluginName'
        - $ref: '#/components/parameters/pluginVersion'
      responses:
        '200':
          description: Plugin metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plugin'
        '404':
          description: Plugin not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      tags: [Plugins]
      summary: Delete plugin
      description: Delete a specific plugin version.
      operationId: deletePlugin
      parameters:
        - $ref: '#/components/parameters/pluginName'
        - $ref: '#/components/parameters/pluginVersion'
      responses:
        '204':
          description: Plugin deleted
        '404':
          description: Plugin not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Plugin is referenced by an artifact
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: Get project
      description: Get a specific project by ID.
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    put:
      tags: [Projects]
      summary: Update project
      description: Update a project.
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      tags: [Projects]
      summary: Delete project
      description: Delete a project and all its resources.
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '204':
          description: Project deleted
        '404':
          description: Project not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # Project Data Planes
...
components:
  parameters:
...
    pluginName:
      name: name
      in: path
      required: true
      description: Plugin name
      schema:
        type: string
    pluginVersion:
      name: version
      in: path
      required: true
      description: Plugin version
      schema:
        type: string
    projectId:
      name: projectId
      in: path
      required: true
      description: Project UUID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Plugin:
      type: object
      required: [name, version, plugin_type, capabilities, config_schema, sha256, registered_at]
      properties:
        name:
          type: string
        version:
          type: string
        plugin_type:
          type: string
          enum: [middleware, dispatcher]
        description:
          type: string
          nullable: true
        capabilities:
          type: array
          items:
            type: string
        config_schema:
          type: object
          description: JSON Schema for plugin configuration
        sha256:
          type: string
          description: SHA256 hash of the WASM binary
        registered_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Error type URI
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Detailed explanation
        instance:
          type: string
          description: URI identifying the specific occurrence
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
    ValidationIssue:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        location:
          type: string
          description: JSON pointer to the error location
...
    Project:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Project name
        description:
          type: string
          nullable: true
          description: Optional project description
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          description: Project name
        description:
          type: string
          nullable: true
          description: Project description
...